Rewrote Friday, Saturday and Sunday classes

+++ Driver.h
@@ -1,16 +1,18 @@
 #ifndef DRIVER_H
 #define DRIVER_H
 
+#include <string>
+
 using namespace std;
 
 class Driver
 {
 	public:
     Driver();
-    Driver(double, double, string);
+    Driver(double, int, string);
     ~Driver();
-    void addPoints(double);
-    double getPoints();
+    void addPoints(int);
+    int getPoints();
     string getName();
     void clearPoints();
     void addSkill(double);
@@ -18,7 +20,7 @@ class Driver
     
   private:
     double skill;
-    double currentPoints;
+    int currentPoints;
     string name;
 };
